+++ veilid-core/src/network_manager/types/events.rs
@@ -1,11 +1,6 @@
 use super::*;
 
-pub(crate) struct PeerInfoChangeEvent {
-    pub routing_domain: RoutingDomain,
-    pub opt_peer_info: Option<Arc<PeerInfo>>,
-}
-
-pub(crate) struct SocketAddressChangeEvent {
+pub struct SocketAddressChangeEvent {
     pub routing_domain: RoutingDomain, // the routing domain this flow is over
     pub socket_address: SocketAddress, // the socket address as seen by the remote peer
     pub old_socket_address: Option<SocketAddress>, // the socket address previously for this peer