+++ veilid-core/src/routing_table/types/events.rs
@@ -1,0 +1,7 @@
+use super::*;
+
+pub struct PeerInfoChangeEvent {
+    pub routing_domain: RoutingDomain,
+    pub opt_old_peer_info: Option<Arc<PeerInfo>>,
+    pub opt_new_peer_info: Option<Arc<PeerInfo>>,
+}