+++ veilid-core/src/routing_table/types/mod.rs
@@ -1,6 +1,7 @@
 mod contact_method;
 mod dial_info_detail;
 mod direction;
+mod events;
 #[cfg(feature = "geolocation")]
 mod geolocation_info;
 mod node_info;
@@ -16,6 +17,7 @@ use super::*;
 pub use contact_method::*;
 pub use dial_info_detail::*;
 pub use direction::*;
+pub use events::*;
 #[cfg(feature = "geolocation")]
 pub use geolocation_info::*;
 pub use node_info::*;